refactor(scene-previewer): extract displayLabel/kindShort into node-labels.ts
Moves the two node-presentation helpers out of SceneInspector so they
can be shared by the upcoming animated-track data service without
duplicating logic. SceneInspector now imports from the shared util;
no behavioral change.
